requirements.txt
pillow>=10.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
psutil>=5.9.4
PyYAML>=6.0
httpx>=0.27.0
sqlalchemy>=2.0.0
d3blocks>=1.0.0
html2image>=2.0.0
pandas>=2.0.0
feedparser>=6.0.10
pytest>=8.2.0